requirements.txt
pandas==1.5.3
pandasai==2.0.42
pyyaml
python-dotenv